Docker-Cloud-Fog-Edge/apps/requirements-python-service.txt
boto3==1.34.131
fastapi==0.115.0
kafka-python==2.0.2
mlflow==2.7.1
numpy==1.26.4
paho-mqtt==1.6.1
pandas==2.2.2
prometheus-client==0.20.0
psycopg2-binary==2.9.9
requests==2.32.3
scikit-learn==1.4.2
scipy==1.13.1
uvicorn==0.30.6
